+++ lib/Braintree/Result/CreditCardVerification.php
@@ -2,6 +2,7 @@
 namespace Braintree\Result;
 
 use Braintree\RiskData;
+use Braintree\ThreeDSecureInfo;
 use Braintree\Util;
 
 /**
@@ -61,6 +62,9 @@ class CreditCardVerification
             $attributes['riskData'] = RiskData::factory($attributes['riskData']);
         }
 
+        if(isset($attributes['threeDSecureInfo'])) {
+            $attributes['threeDSecureInfo'] = ThreeDSecureInfo::factory($attributes['threeDSecureInfo']);
+        }
         $this->_attributes = $attributes;
         foreach($attributes AS $name => $value) {
             $varName = "_$name";